move specialized DB functions into services folder
